Adds a dividing line
Most of this installation is user based, allowing for users to maintain
different versions of cpan and duckpan. For starters we will mimic that
setup in our Dockerfile, however er may want to later revisit if this is
necessary/a good idea. Since Docker creates small dispensable machines
meant for rapid deployment and testing, perhaps we could move all this
to root.