This is an alternative
This method installs everything as root. I would want more input from
the duckpan team as to why the install install script refuses to install
under root.

What I like about this way is it gets the PATH variable all set so that
perl --version returns the expected version, so I may end up taking
pieces of this for the final script.